Develop

+++ tests/main.py
@@ -1,31 +1,38 @@
 import sys
 import os
+import enum
 
 from pathlib import Path
-from utils.blender import b3d_launch_blender_test
-from utils.container import clear_container_test
+from utils.blender import get_b3d_addon_dependency
+from utils.container import VirtualMachine
 from utils.misc import ordering_test_file, generate_archive
+from utils.properties import ContainerObject
 
 
-def launch_unit_test(tag: str, test: [str] = None):
+class Container(enum.Enum):
+    """Enumerate about the Geometry node"""
+    BLENDER = ContainerObject(name='Blender', image='stilobique/blender')
+
+
+def launch_unit_test(test: [str] = None):
     """Start all Unit Test, Blender and Unreal if needed"""
     # Ordering Unit test Blender/Unreal
     if test is None:
         test = ordering_test_file()
-        test = test['blender']
+        b3d = test['blender']
+    else:
+        b3d = test
 
-    b3d_launch_blender_test(test=test, tag=tag)
+    vm_b3d = VirtualMachine(Container.BLENDER.value)
+    vm_b3d.launch_unit_test(tests=b3d)
 
 
 if __name__ == '__main__':
     # Initialize Variable and module request
     archives = []
-    docker_tag = 'stilobique/blender:latest'
-
-    # Clear blender container
-    clear_container_test(tag=docker_tag)
 
     # Prepare Blender and Unreal dependency
+    get_b3d_addon_dependency(archives)
     generate_archive(archives, 'blender_addon_folder')
 
     # Generate Unit Test, a specific call or execute all Unit Test
@@ -40,8 +47,9 @@ if __name__ == '__main__':
                     test_list.append(item)
 
     # Launch Unit Test
-    launch_unit_test(test=test_list, tag=docker_tag)
+    launch_unit_test(test=test_list)
 
     # Clear archive file and container
     for archive in archives:
-        os.remove(Path(os.getcwd(), archive))
+        if Path(Path(os.getcwd(), archive)).exists():
+            os.remove(Path(os.getcwd(), archive))
